@@ -1,74 +1,69 @@
-﻿using Microsoft.AspNetCore.Http;
-using Microsoft.Extensions.Logging;
-using System.Collections.Generic;
-using System.Linq;
-using System.Reflection;
+﻿using System.Reflection;
 
-namespace HttpServerSample
+namespace HttpServerSample;
+
+public class GenerateHtml
 {
-    public class GenerateHtml
+    private readonly ILogger _logger;
+    public GenerateHtml(ILogger<GenerateHtml> logger)
     {
-        private readonly ILogger _logger;
-        public GenerateHtml(ILogger<GenerateHtml> logger)
+        _logger = logger;
+    }
+
+    private static string s_htmlFormat =
+        "<!DOCTYPE html>\r\n<html><head><title>{0}</title></head>" +
+        "<body>{1}</body></html>";
+
+    public string GetHtmlContent(HttpRequest request)
+    {
+        string title = "Sample Listener using Kestrel";
+
+        string content = $"<h1>Hello from the server</h1>" +
+            $"<h2>Header Info</h2>" +
+            $"{string.Join(' ', GetHeaderInfo(request.Headers))}" +
+            $"<h2>Request Object Information</h2>" +
+            $"{string.Join(' ', GetRequestInfo(request))}";
+
+        return string.Format(s_htmlFormat, title, content);
+    }
+
+    private IEnumerable<string> GetRequestInfo(HttpRequest request)
+    {
+        var properties = request.GetType().GetProperties();
+        List<(string Key, string Value)> values = new();
+        foreach (var property in properties)
         {
-            _logger = logger;
-        }
-
-        private static string s_htmlFormat =
-            "<!DOCTYPE html>\r\n<html><head><title>{0}</title></head>" +
-            "<body>{1}</body></html>";
-
-        public string GetHtmlContent(HttpRequest request)
-        {
-            string title = "Sample Listener using Kestrel";
-
-            string content = $"<h1>Hello from the server</h1>" +
-                $"<h2>Header Info</h2>" +
-                $"{string.Join(' ', GetHeaderInfo(request.Headers))}" +
-                $"<h2>Request Object Information</h2>" +
-                $"{string.Join(' ', GetRequestInfo(request))}";
-
-            return string.Format(s_htmlFormat, title, content);
-        }
-
-        private IEnumerable<string> GetRequestInfo(HttpRequest request)
-        {
-            var properties = request.GetType().GetProperties();
-            List<(string Key, string Value)> values = new();
-            foreach (var property in properties)
+            try
             {
-                try
+                string? value = property.GetValue(request)?.ToString();
+                if (value != null)
                 {
-                    string? value = property.GetValue(request)?.ToString();
-                    if (value != null)
-                    {
-                        values.Add((property.Name, value));
-                    }
-                }
-                catch (TargetInvocationException ex)
-                {
-                    _logger.LogInformation("property name: {0}: message {1}", property.Name, ex.Message);
-                    if (ex.InnerException != null)
-                    {
-                        _logger.LogInformation("\tinner exception: {0}", ex.InnerException.Message);
-                    }
+                    values.Add((property.Name, value));
                 }
             }
-            return values.Select(v => $"<div>{v.Key}: {v.Value}</div>");
-        }
-
-        private IEnumerable<string> GetHeaderInfo(IHeaderDictionary headers)
-        {
-            List<(string Key, string Value)> values = new();
-            var keys = headers.Keys;
-            foreach (var key in keys)
+            catch (TargetInvocationException ex)
             {
-                if (headers.TryGetValue(key, out var value))
+                _logger.LogInformation("property name: {0}: message {1}", property.Name, ex.Message);
+                if (ex.InnerException != null)
                 {
-                    values.Add((key, value));
+                    _logger.LogInformation("\tinner exception: {0}", ex.InnerException.Message);
                 }
             }
-            return values.Select(v => $"<div>{v.Key}: {v.Value}</div>");
         }
+        return values.Select(v => $"<div>{v.Key}: {v.Value}</div>");
+    }
+
+    private IEnumerable<string> GetHeaderInfo(IHeaderDictionary headers)
+    {
+        List<(string Key, string Value)> values = new();
+        var keys = headers.Keys;
+        foreach (var key in keys)
+        {
+            if (headers.TryGetValue(key, out var value))
+            {
+                values.Add((key, value));
+            }
+        }
+        return values.Select(v => $"<div>{v.Key}: {v.Value}</div>");
     }
 }
