@@ -1,5 +1,4 @@
-﻿using System;
-using System.Net.Http.Headers;
+﻿using System.Net.Http.Headers;
 
 static class Utilities
 {
@@ -14,4 +13,3 @@ static class Utilities
         Console.WriteLine();
     }
 }
-
